rename: add dto/user package and move files
