+++ src/main/java/com/example/backend/controller/UserController.java
@@ -3,6 +3,9 @@ package com.example.backend.controller;
 import com.example.backend.domain.User;
 import com.example.backend.dto.*;
 import com.example.backend.dto.login.KaKaoAuthRequestDTO;
+import com.example.backend.dto.user.UserJoinRequestDTO;
+import com.example.backend.dto.user.UserModifyRequestDTO;
+import com.example.backend.dto.user.UserProfileDTO;
 import com.example.backend.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
