+++ src/main/java/com/example/backend/dto/user/UserModifyRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.backend.dto;
+package com.example.backend.dto.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
