+++ src/main/java/com/example/backend/service/UserServiceImpl.java
@@ -2,8 +2,8 @@ package com.example.backend.service;
 
 import com.example.backend.domain.User;
 import com.example.backend.dto.login.KaKaoAuthRequestDTO;
-import com.example.backend.dto.UserJoinRequestDTO;
-import com.example.backend.dto.UserModifyRequestDTO;
+import com.example.backend.dto.user.UserJoinRequestDTO;
+import com.example.backend.dto.user.UserModifyRequestDTO;
 import com.example.backend.global.exception.InvalidInputException;
 import com.example.backend.global.exception.InvalidInputExceptionType;
 import com.example.backend.repository.UserRepository;
